tpm: Implement ioctl to set a tpm_chip on an IMA namespace
Implement an ioctl for the vtpm proxy 'server side'. The ioctl is
available for the anonymous file descriptor returned from the operation
on /dev/vtpmx. It enables the transfer of a tpm_chip to a user_namespace
and the setting of a tpm_chip on an ima_namespace.
The operation may be rejected if a tpm_chip is already in use by an IMA
namespace or the IMA namespace already has a tpm_chip or if IMA
namespacing is not supported.

The tpm_chip is later on released when the IMA namespace is torn down.
The teardown may take a while since user namespace teardown is only
occurring much later than for example the termination of a user namespace.
Therefore, seemingly stale TPM devices (/dev/tpmX and /dev/tpmrmX) may be
kept around for a while until the user namespace is freed.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>