ima: Move ima_init_digests() into init_ima_namespace()
Move the call to ima_init_digests() from ima_init() into
init_ima_namespace() and preserve the order of initialization by
calling it at the end of init_ima_namespace(). Call this function
for when any ima_namespace is initialized.

Implement ima_free_digests() to free the ima_digests structure
of an ima_namespace when the ima_namespace is freed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>