ima: Define ima_namespace struct and start moving variables into it
Define the ima_namespace structure and the ima_namespace variable
init_ima_ns for the host's IMA namespace. Implement basic functions for
namespacing support.

Move variables related to the IMA policy into the ima_namespace. This way
the IMA policy of an IMA namespace can be set and displayed using a
front-end like securityfs.

Implement ima_ns_from_file() to get the IMA namespace via the user
namespace of the securityfs superblock that a file belongs to.

To get the current ima_namespace use &init_ima_ns when a function
that is related to a policy rule is called.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>
Acked-by: Christian Brauner <brauner@kernel.org>

---

v9:
 - squashed patched 2 and 3 of v8
 - ima_post_read_file: only access ima_appraise in case of init_ima_ns