integrity: Move keyrings into integrity_namespace
Define an integrity_namespace structure and move the keyrings into it.
Also define an init_integrity_ns that holds the hosts keyrings.

Create the integrity_namespace when creating the user_namespace.
The reason for this is that runc causes a call to evm_inode_setattr()
soon after creating a user namespace. The get_intgerity_ns() there must
not result in NULL pointer. The integrity_namespace needs access to
keyrings (_evm, _ima) that may be there independent of SecurityFS or an
IMA namespace.

Reference the init_integrity_ns from the init_ima_ns.

Pass a pointer to the integrity_namespace structure through the APIs
so that the keyrings can be accessed through this structure now.

Define current_integrity_ns() for configurations when CONFIG_IMA_NS is
defined or not defined.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>