security/ima: Enable setting security.ima in IMA namespace
Modify cap_inode_setxattr() and cap_inode_removexattr() so that the
evaluation of capabilities required to access security.ima is deferred to
ima_inode_setxattr() and ima_inode_removexattr() respectively if
CONFIG_IMA_NS is enabled as well as CONFIG_IMA_APPRAISE (avoid the
simple stub function)

Modify ima_protect_xattr(), which is called by ima_inode_setxattr() and
ima_inode_removexattr(), so that the capabilities are evaluated there.
Add the ns_capable() check previously executed by cap_inode_setxattr() to
this function. Further, for allowing setting the IMA xattr inside an IMA
namespace, also allow to use CAP_MAC_ADMIN so that regular users spawning
an IMA namespace can also sign files in the IMA namespace if those files
have owner and group mappings in the user namespace.
If both CAP_SYS_ADMIN and CAP_MAC_ADMIN are missing in an IMA namespace,
or of the file owner and group cannot be mapped then writing and removing
of security.ima xattr is not possible.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>