ima: Adjust flags in loop over ns_status when clearing atomic bits
Rather than only clearing flags on the iint, walk the list of ns_status
connected to the iint and clear the flags on each ns_status. This is a
necessary step to prepare for moving appraisal-related flags from the iint
into the ns_status so they can then be handled on a per-namespace basis.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>