ima: Clear the IMA_UPDATE_XATTR bit if none of the IMA-ns must appraise
Adjust the implementation of ima_inode_post_setattr() for the namespacing
case. Where previously looking at the iint was sufficient, we now have
to consider the case of shared files and therefore walk the list of
ns_status and check for an active appraise policy on the IMA namespaces
found in the list of ns_status. Call ima_must_appraise() for each IMA
namespace with an appraise policy terminate the loop once the 'action'
was found because in this case the IMA_UPDATE_XATTR will not have to be
reset.

Reset the IMA_UPDATE_XATTR flag only of any namespaces with an appraise
policy were found but none of them had the 'action' returned from the
ima_must_appraise() call.

Since the ns_status now always needs access to the ima_namespace, make
the pointer also accessible if CONFIG_IMA_NS is not set.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>