ima: Enable configuration of hash algo using securityfs 'hash' file
Enable the configuration of the hash algo IMA will be using a new
'hash' file where a user may write the name of the hash into in the IMA
namespace configuration step. Once the IMA namespace is active, the
file can be read and shows the hash algorithm that is used.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>