ima: Move ima_tpm_chip and crypto related structures into ima_namespace
Move the ima_tpm_chip and a first set of crypto related structures into
ima_namespace. The crypto related structure need to be moved since they
depend on the number of PCR banks the ima_tpm_chip has. This then allows us
to namespace ima_init_crypto() where the initialization of the algorithms
depends on the TPM's number of PCR banks and where a NULL pointer for
the tpm_chip is already handled correctly as needed for the IMA namespacing
case where an IMA namespace will not have an associate tpm_chip initially.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>