ima: Pass optional template name to ima_init_template
Pass an optional template name to ima_init_template(), which then passes
the name onto ima_template_desc_current(). Call ima_init_template() with
the template name from the namespace's configuration and call it with the
NULL parameter in case of the init_ima_ns, which then results in activation
of the default template as before.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>