integrity/ima: Move ima_hash from iint into ns_status
Move the ima_hash field from the iint into the ns_status data structure
so that the ima_digest_data structure can now be created in
ima_collect_measurement() on a per IMA namespace basis.

Also move the IMA_COLLECTED flag from the iint flags into the ns_status
structure since this flag is tightly coupled with the ima_hash field and
it needs to now be set on a per-IMA namespace basis.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>