ima: Add support for UUID field to template fields and library
Add support for a UUID field to the template fields. Use the UUID of the
user namespace that caused an event for logging the event in all parent
IMA namespaces' logs.

Pass the uuid of the source user namespace where an event occurred as part
of the event_data in the uuid field.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>