evm: Copy evm_config_default_xattrnames before appending to list
Make a copy of each element in evm_config_deault_xattrs before
appending it to a list so that each element can be appended
to a namespace-specific list later on. Extend the xattr_list
structure with a boolean indicating whether the name was allocated
and needs to be freed. Static strings have allocated = false.

Implement function to free an xattr_list struct and a function
to free a list of xattr_lists.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>