evm: Move call to evm_init_config into evm_init_namespace
Move the call to evm_init_config() into evm_init_namespace() so that this
function is now called for the init_evm_ns and all EVM namespaces as
well.

Restrict the display of info using pr_info to only be print when the
init_evm_ns is initialized. Since init_evm_ns and EVM namespaces share the
same data this displays the same information as EVM namespaces would.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>