evm: Enable setting security.evm in EVM namespace
Modify cap_inode_setxattr() and cap_inode_removexattr() so that the
evaluation of capabilities required to access security.evm is deferred to
ima_inode_setxattr() and ima_inode_removexattr() respectively.

Modify evm_protect_xattr(), which is called by evm_inode_setxattr() and
evm_inode_removexattr(), so that the capabilities are evaluated there.
Add the ns_capable() check previously executed by cap_inode_setxattr() to
this function. Further, for allowing setting the EVM xattr inside an EVM
namespace, also allow to use CAP_MAC_ADMIN so that regular users spawning
an EVM namespace can also sign files in the EVM namespace if those files
have owner and group mappings in the user namespace.
If both CAP_SYS_ADMIN and CAP_MAC_ADMIN are missing in an EVM namespace,
or if the file owner and group cannot be mapped then writing and removing
of security.evm xattr is not possible.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>