security: Pass security.* xattrs on to EVM if CAP_MAC_ADMIN is set
Allow the handling of all security.* xattrs in user namespaces by
EVM if CAP_MAC_ADMIN is set and the inode is fully owned by the
namespace. security.* xattr can now be altered in a user namespace
if a file is fully mapped (uid and gid) in the user namespace and
if EVM is enabled (on top of IMA namespacing being enabled) and the
user has CAP_MAC_ADMIN in the user namespace.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>