fs/xattr: Add check_flags to vfs_setxattr to adjust behavior
To be able to influence the behavior of EVM's setxattr implementation and
have it avoid the check whether the current security.evm xattr validates,
add a check_flag to vfs_setxattr to be able to pass this flag. Overlayfs
will later on use this flag when setting xattrs on upper files that would
not be able to validate security.evm since they have not been set there.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>