crypto: ecc - Implement and use ecc_curve_get_nbytes to get curve's nbytes
Implement ecc_curve_get_nbytes to get a curve's number of bytes (nbytes).
The number of bytes is derived from the nbits field of a curve. This
function returns a precise result for the number of bytes. This is needed
for NIST P521 where nbytes cannot be derived from ndigits.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>