evm: Store metadata inode attributes and detect changes
The inode holding the metadata on stacked filesystems may be different
than the one holding the file data and therefore changes to it need to be
detected independently. Therefore, store the i_version, device number, and
inode number associated with the file metadata inode.

Implement a function to detect changes to the inode and if a change is
detected reset the evm_status. This function will be called by IMA when
IMA detects that the metadata inode is different from the one holding
the file data and the one representing the file.

Co-developed-by: Mimi Zohar <zohar@linux.ibm.com>
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>