tpm: of: If available use linux,sml-log to get the log and its size
If linux,sml-log is available use it to get the TPM log rather than the
pointer found in linux,sml-base. This resolves an issue on PowerVM and KVM
on Power where after a kexec the memory pointed to by linux,sml-base may
have become inaccessible or corrupted. This was due to the fact that the
buffer was not properly protected across the kexec. Also, linux,sml-log
has replaced linux,sml-base and linux,sml-size on these two platforms.

Keep the handling of linux,sml-base/sml-size for PowerNV and other
platforms, that may provide the two properties via firmware, such as
skiboot in case of PowerNV.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>