powerpc/prom_init: Replace linux,sml-base/sml-size with linux,sml-log
linux,sml-base holds the address of a buffer with the TPM log. This buffer
may become invalid due to a kexec since it was not properly protected. To
avoid accessing an invalid address or corrupted buffer, embed the whole TPM
log in the device tree using the property linux,sml-log. This helps to
protect the log since the device tree is protected across a kexec soft
reboot with both of the kexec syscalls.

Avoid having the firmware ingest the whole TPM log when calling
prom_setprop but initially only create the linux,sml-log property as a
place holder. Insert the actual TPM log during the tree flattening phase.

Suggested-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>